refactor: memperbaiki struktur folder dan penamaannya
struktur folder dengan menggunakan atomic, penamaan dengan menggunakan camel case
