+++ README.md
@@ -1,3 +1,5 @@
 running 
 
 npm run dev
+
+camelcase
