--- src/Component/home/ui/coba.tsx
@@ -1,8 +1,0 @@
-function Coba () {
-  return (
-    <div>
-      <h1>Coba</h1>
-    </div>
-  )
-}
-export default Coba;
