+++ src/components/Card/card_blog.tsx
