+++ src/components/Card/card_features.tsx
