+++ src/pages/Home/index.tsx
@@ -1,0 +1,13 @@
+import Hero from "./hero";
+
+function Home() {
+  return (
+    <>
+      <div className="h-full w-full">
+        <Hero />
+      </div>
+    </>
+  );
+}
+
+export default Home;
