+++ src/pages/Home/testimoni.tsx
@@ -1,4 +1,4 @@
-import SectionHeader from "./ui/header";
+import SectionHeader from "../elements/header/header";
 import star from "../../assets/svg/testimony/Star.svg";
 import SectionButton from "./ui/button";
 import SectionCardTestimony from "./ui/card_testimony";
